+++ packages/powersync_core/lib/powersync_core.dart
@@ -9,7 +9,6 @@ export 'src/database/powersync_database.dart';
 export 'src/exceptions.dart';
 export 'src/log.dart';
 export 'src/open_factory.dart';
-export 'src/powersync_database.dart';
 export 'src/schema.dart';
 export 'src/sync_status.dart';
 export 'src/uuid.dart';